fix ossl api initialization race
This is a bit of a hack

ossl_namemap_stored has this race in which it calls
ossl_lib_ctx_get_data to get the context namemap, and then, internal to
ossl_namemap_stored, populates the table.  This is fine in the single
threaded case, as there is only one thread doing the population and
lookup operations.  In the multithreaded case however, its possible for
two threads to race her, in which Thread2 gets the namemap and attempts
to preform a lookup while Thread1 is still populating the hash table.

This is fine in and of itself, as lookups can always be retried, but our
test suite (and likely many user applications), assume that, at least
for core names, lookups are synchronous behind the table population.
Thats not really a valid assumption, but I don't think its reasonable to
violate it.

RUN_ONCE isn't helpful here, as the population occurs after RUN_ONCE
completes, and so subsequent readers aren't blocked, and the table
population can't be moved into the RUN_ONCE routine, because it accepts
no parameters, and so we have no way to communicate to the init routine
that what we need to initalize.

So, what we need is a mechanim that:
1) Allows threads to wait, until the initalization of the hash table is
   complete
2) Allows a single thread passing through (1) to advance, so that the
   table population can complete
3) Be as close to zero time cost as possible once the initalization is
   complete

So this is what I've come up with.  An new call
ossl_namemap_wait_init_done that does the following:
a) it aquires an rcu read lock on the namemap
b) checks the namemap ready variable atomically to see if it is NULL or
a pending value (currently 0x1)
c) If it is pending, drop the readlock and go to (a)
d) if it is NULL, drop the read lock and acquire the rcu write lock for
the namemap.
e) check the ready variable again, if it is set to pending, drop the
write lock and go to (a)
f) If it is NULL, assign the pending value to the namemap ready variable
and break the loop

Later, after a thread that hits point (f) above finishes
ossl_namemap_stored, it updates the ready variable to a complete state
(currently just pointing at its own struct

The effect here is an only-one mechanism, in which whatever thread gets
to the write lock first populates the namemap table, while other threads
spin waiting for completion.  Once this startup operation is complete,
the wait function effectively reduces to an
rcu_read_lock/deref/rcu_unlock, which should be very fast, and that will
be the performance profile for the lifespan of the process from that
point forward

This is a bit rough, If there are other use cases where this is needed,
this should be packaged into a macro the thread_once.h header I think