convert ADDED_OBJ hash table to be refcounted
Note: This took some extra alteration.  The ADDED_OBJ hash table, on
add, actually adds 3 entries to the table, each pointing to the same
asn1 object.  This previously required some odd gymnastics on free, in
which the nid value was used as a temporary reference counter, which was
set to zero, incremented, and decremented in 3 separate doall loops.
Discard that, and instead just allocate a single wrapper structure on
add with a usage counter and point each of the 3 inserts to it.  That
way we can just drop the refcount on free, and free the asn1 obj when
the counter hits zero