Accomodate NonStop definition of sockaddr_storage
Currently, on NonStop, sockaddr_storage appears to be far too small to
accomodate struct sockaddr_un (the former being about 198 bytes, the
latter 1040 bytes)

This makes the use of sockaddr_storage impossible for AF_UNIX sockets in
which the sun_path address is more than 182 bytes long, which it is in
our bio_addr_test.

Accomodate that by switching to a union of explicit sockaddr types, like
we do in make_dummy_addr to ensure that we have sufficient automatic
storage to hold the full 1024 bytes that NonStop defines as the length
of sun_path.

Fixes #22218