Create an rcu locked/refcounted variant of the hash tables
One of the instanced in which we use rwlocks extensively is in
management of our hash tables.  Create a hash table variant that is
internally locked using rcu locks to increase performance.

Because the hash table is locked internally to the api, we need to
account for the fact that we would have to unlock the rcu lock before
returning data that we read/write/delete.  To accomodate this, make the
data entries in the table refcounted, and add appropriate apis to drop
said refcount, and manage freeing