Make chrome recognize files as needing to be downloaded
Nominally, chrome treats text/html content as something to display in
the browser.  here though we want to ensure that content is downloaded
for interop tests.

Do the following:

Set content-disposition header to be attachment - optional, but helps
chrome decide that this is a downloadable file

set content-type header to application/octet-stream - required for
chrome to be able to download files

set content-length, optional, but it helps identify files in tcpdumps if
we're debugging something