Reduce quic default MTU to 1200 bytes.
The MTU size is somewhat convoluted with the minimum quic packet size
here.  The smallest allowable QUIC frame is 1200 bytes, not 1500, so
we're generally allocating 300 bytes extra per urxe when we're not sure
its used.

Lowering the size also has the advantage for forcing a pass through
ossl_quic_resize_urxe in the event that our actual mtu is higher, and
QUIC sends a packet that is greater than the 1200 byte minimum.

This has the advantage of exercising our resize path for urxes,
increasing our test coverage.