Fixup qtx_resize_txe and make it a common path for allocation
qtx_resize_txe assumes that the txl list passed in will never be NULL,
but it is passed as NULL from qtx_write.  We just never hit that path
because allocated qtx objects are always of sufficient size to never
need resizing.

Adjust qtx_resize_txe to treat a NULL txl as meaning an object has
already been removed from the list, so not list adjustment is needed.

Also, use qtx_resize_txe in the qtx_alloc_txe path to ensure that this
code gets some testing exercise.