Replace use of big write lock in namemap with a lockless linked list
Instead of just always taking the write lock on the addition of a name
or names, use a lockless linked list to to create a waiting filter.

Each name add request adds its names to the pending list with a
monotonically incrementing identifier.

After addition each proposed add request iterates over the list looking
for previously added requests that match at least one name in this
requests list.  if a match is found, iteration continues, until no
matches are found.

When no matches are found the request procededs adding its names to the
namemap using either a unique new number of the number allocated for a
previously matching name.

When the addition is complete, the request removes its entry from the
linked list, allowing subsequent additions to procede.

This still creates delay in namemap additions, but only for those
requests with matching names, independent names can procede in parallel