sort the X509_OBJECT list only when we modify it.
When we call get_cert_by_subject_ex we always take the store lock at the
end of the function to potentially sort the X509_OBJECT list.  This is
unneeded, as we may only rarely add to the object store.  Instead only
do the sort on an actuall push to the stack in add_dir.