change stack of BY_DIR_HASH to lhash
The BY_DIR_HASH stack gets sorted alot when we call
get_cert_by_subject_ex.  We frequently add entries to this list,
requiring an expensive sort while holding the write lock.  Since this
struct is sorted by a hash value, it seems natural that we store it in a
hash table instead, obviating the need to sort it on every update.