hqc encoder/decoder implementation
NOTE: This commit introduces a OID for HQC in the OpenSSL PEN
namespace, which is very wrong, but needed at the moment until an OID is
allocated by ITU/IANA for the algorithm.  This also encodes keys using
our standard encoders, which is almost certainly wrong.  When the
standard gets fleshed out more we will likely need to drop this and
replace it with the correct encoding format, or drop it entirely if
these keys are just meant to be ephemeral.

Its really just here right now to help people experiment with it.