format generated files
We have a large number of files that get generated with make update, and
then those generated files are comitted to the git history.

That means we are left with an upleasant choice: either ignore them for
the purpose of formatting, or alter the generator scripts to attempt to
match the expected clang style when generating them.

This branch seeks a middle ground, in which the make update target,
after generating the files, runs clang-format on them, so that post
generation, we run them through style formatting prior to commit.

Its not ideal, as it creates a dependency on having clang-format
installed locally (though we might could mitigate that by optionally
doing the formatting based on a presence check of the format binary,
leaving users without the tool to rely on the CI generated patch to fix
it up after PR creation)