disable formatting for obj_dat.h
obj_dat.h is a generated file that requires formatting, generally
speaking, but we have a few arrays in there that are very specifically
formatted (the so and nid_objs arrays).  We take care to place certain
sets of values on a single line with comments, so is to differentiate
data elements that are associated, but not reflected by the type of the
array.  As such, disable formatting for them